+++ arg/arg_test.go
@@ -44,22 +44,22 @@ func (s *ArgUtilSuite) TestAdd(c *C) {
 func (s *ArgUtilSuite) TestAddMap(c *C) {
 	args := NewArguments()
 
-	m1 := Map{"": &V{}}
-	m2 := Map{"t:": &V{}}
+	m1 := Map{"": {}}
+	m2 := Map{"t:": {}}
 
 	c.Assert(args.AddMap(m1), Not(HasLen), 0)
 	c.Assert(args.AddMap(m2), Not(HasLen), 0)
 
-	m3 := Map{"test": &V{}}
-	m4 := Map{"t:test2": &V{}}
+	m3 := Map{"test": {}}
+	m4 := Map{"t:test2": {}}
 
 	c.Assert(args.AddMap(m3), HasLen, 0)
 	c.Assert(args.AddMap(m4), HasLen, 0)
 
 	m5 := Map{
-		"t:test":  &V{},
-		"T:test1": &V{},
-		"t:test2": &V{},
+		"t:test":  {},
+		"T:test1": {},
+		"t:test2": {},
 	}
 
 	c.Assert(args.AddMap(m5), HasLen, 2)
@@ -92,10 +92,10 @@ func (s *ArgUtilSuite) TestGlobal(c *C) {
 	_, errs := global.Parse(
 		strings.Split("-s Test -i 123 -f 100.5 -b", " "),
 		Map{
-			"s:string": &V{},
-			"i:int":    &V{Type: INT},
-			"f:float":  &V{Type: FLOAT},
-			"b:bool":   &V{Type: BOOL},
+			"s:string": {},
+			"i:int":    {Type: INT},
+			"f:float":  {Type: FLOAT},
+			"b:bool":   {Type: BOOL},
 		},
 	)
 
@@ -113,12 +113,12 @@ func (s *ArgUtilSuite) TestLimiters(c *C) {
 	argline := "--int1 1 --int2 5 --int3 10 --float1 1.0 --float2 5.0 --float3 10.0"
 
 	argsMap := Map{
-		"int1":   &V{Type: INT, Min: 3, Max: 7},
-		"int2":   &V{Type: INT, Min: 3, Max: 7},
-		"int3":   &V{Type: INT, Min: 3, Max: 7},
-		"float1": &V{Type: FLOAT, Min: 3.0, Max: 7.0},
-		"float2": &V{Type: FLOAT, Min: 3.0, Max: 7.0},
-		"float3": &V{Type: FLOAT, Min: 3.0, Max: 7.0},
+		"int1":   {Type: INT, Min: 3, Max: 7},
+		"int2":   {Type: INT, Min: 3, Max: 7},
+		"int3":   {Type: INT, Min: 3, Max: 7},
+		"float1": {Type: FLOAT, Min: 3.0, Max: 7.0},
+		"float2": {Type: FLOAT, Min: 3.0, Max: 7.0},
+		"float3": {Type: FLOAT, Min: 3.0, Max: 7.0},
 	}
 
 	args := NewArguments()
@@ -132,29 +132,61 @@ func (s *ArgUtilSuite) TestLimiters(c *C) {
 	c.Assert(args.GetF("float3"), Equals, 7.0)
 }
 
+func (s *ArgUtilSuite) TestConflicts(c *C) {
+	argline := "--test1 abc --test2 123"
+
+	argsMap := Map{
+		"test1": {Conflicts: "test2"},
+		"test2": {},
+	}
+
+	args := NewArguments()
+	_, errs := args.Parse(strings.Split(argline, " "), argsMap)
+
+	c.Assert(errs, Not(HasLen), 0)
+	c.Assert(errs[0].(ArgumentError).Type, Equals, ERROR_CONFLICT)
+	c.Assert(errs[0].Error(), Equals, "Argument test1 conflicts with argument test2")
+}
+
+func (s *ArgUtilSuite) TestBound(c *C) {
+	argline := "--test1 abc"
+
+	argsMap := Map{
+		"test1": {Bound: "test2"},
+		"test2": {},
+	}
+
+	args := NewArguments()
+	_, errs := args.Parse(strings.Split(argline, " "), argsMap)
+
+	c.Assert(errs, Not(HasLen), 0)
+	c.Assert(errs[0].(ArgumentError).Type, Equals, ERROR_BOUND_NOT_SET)
+	c.Assert(errs[0].Error(), Equals, "Argument test2 must be defined with argument test1")
+}
+
 func (s *ArgUtilSuite) TestGetters(c *C) {
 	argline := "file.mp3 -s STRING --required TEST -i 320 -b -f 1.098765 -S2 100 -f1 5 -f2 1 -ms ABC --merg-string DEF -mi 6 --merg-int 6 -f3 12 -mf 10.1 -mf 10.1 -i1 5"
 
 	argsMap := Map{
-		"s:string":          &V{Type: STRING, Value: "STRING"},
-		"S:empty-string":    &V{Type: STRING},
-		"r:required":        &V{Required: true, Alias: "A:alias"},
-		"i:int":             &V{Type: INT},
-		"i1:int-between":    &V{Type: INT, Min: 1, Max: 3},
-		"I:not-set-int":     &V{Type: INT, Value: 0},
-		"b:bool":            &V{Type: BOOL},
-		"B:empty-bool":      &V{Type: BOOL},
-		"B1:not-set-bool":   &V{Type: BOOL, Value: false},
-		"f:float":           &V{Type: FLOAT},
-		"F:not-set-float":   &V{Type: FLOAT, Value: 0.0},
-		"f1:float-max":      &V{Type: FLOAT, Max: 3.0},
-		"f2:float-min":      &V{Type: FLOAT, Min: 3.0},
-		"f3:float-between":  &V{Type: FLOAT, Min: 3.0, Max: 10.0},
-		"ms:merg-string":    &V{Mergeble: true},
-		"mi:merg-int":       &V{Type: INT, Mergeble: true},
-		"mf:merg-float":     &V{Type: FLOAT, Mergeble: true},
-		"S1:not-set-string": &V{Type: STRING, Value: ""},
-		"S2:string-as-num":  &V{Type: STRING},
+		"s:string":          {Type: STRING, Value: "STRING"},
+		"S:empty-string":    {Type: STRING},
+		"r:required":        {Required: true, Alias: "A:alias"},
+		"i:int":             {Type: INT},
+		"i1:int-between":    {Type: INT, Min: 1, Max: 3},
+		"I:not-set-int":     {Type: INT, Value: 0},
+		"b:bool":            {Type: BOOL},
+		"B:empty-bool":      {Type: BOOL},
+		"B1:not-set-bool":   {Type: BOOL, Value: false},
+		"f:float":           {Type: FLOAT},
+		"F:not-set-float":   {Type: FLOAT, Value: 0.0},
+		"f1:float-max":      {Type: FLOAT, Max: 3.0},
+		"f2:float-min":      {Type: FLOAT, Min: 3.0},
+		"f3:float-between":  {Type: FLOAT, Min: 3.0, Max: 10.0},
+		"ms:merg-string":    {Mergeble: true},
+		"mi:merg-int":       {Type: INT, Mergeble: true},
+		"mf:merg-float":     {Type: FLOAT, Mergeble: true},
+		"S1:not-set-string": {Type: STRING, Value: ""},
+		"S2:string-as-num":  {Type: STRING},
 	}
 
 	args := NewArguments()
@@ -240,91 +272,91 @@ func (s *ArgUtilSuite) TestParsing(c *C) {
 
 	// //////////////////////////////////////////////////////////////////////////////// //
 
-	_, errs = NewArguments().Parse([]string{"--test", "100"}, Map{"t:test": &V{Type: 10}})
+	_, errs = NewArguments().Parse([]string{"--test", "100"}, Map{"t:test": {Type: 10}})
 
 	c.Assert(errs, Not(HasLen), 0)
 	c.Assert(errs[0].Error(), Equals, "Unsuported argument type 10")
 
 	// //////////////////////////////////////////////////////////////////////////////// //
 
-	_, errs = NewArguments().Parse([]string{}, Map{"s:string": &V{}, "s:trace": &V{}})
+	_, errs = NewArguments().Parse([]string{}, Map{"s:string": {}, "s:trace": {}})
 
 	c.Assert(errs, Not(HasLen), 0)
 	c.Assert(errs[0].Error(), Equals, "Argument -s defined 2 or more times")
 
 	// //////////////////////////////////////////////////////////////////////////////// //
 
-	_, errs = NewArguments().Parse([]string{"-t"}, Map{"s:string": &V{}})
+	_, errs = NewArguments().Parse([]string{"-t"}, Map{"s:string": {}})
 
 	c.Assert(errs, Not(HasLen), 0)
 	c.Assert(errs[0].Error(), Equals, "Argument -t is not supported")
 
 	// //////////////////////////////////////////////////////////////////////////////// //
 
-	_, errs = NewArguments().Parse([]string{"-t=100"}, Map{"s:string": &V{}})
+	_, errs = NewArguments().Parse([]string{"-t=100"}, Map{"s:string": {}})
 
 	c.Assert(errs, Not(HasLen), 0)
 	c.Assert(errs[0].Error(), Equals, "Argument -t is not supported")
 
 	// //////////////////////////////////////////////////////////////////////////////// //
 
-	_, errs = NewArguments().Parse([]string{"--test"}, Map{"s:string": &V{}})
+	_, errs = NewArguments().Parse([]string{"--test"}, Map{"s:string": {}})
 
 	c.Assert(errs, Not(HasLen), 0)
 	c.Assert(errs[0].Error(), Equals, "Argument --test is not supported")
 
 	// //////////////////////////////////////////////////////////////////////////////// //
 
-	fArgs, errs := NewArguments().Parse([]string{"-", "--"}, Map{"t:test": &V{}})
+	fArgs, errs := NewArguments().Parse([]string{"-", "--"}, Map{"t:test": {}})
 
 	c.Assert(errs, HasLen, 0)
 	c.Assert(fArgs, DeepEquals, []string{"-", "--"})
 
 	// //////////////////////////////////////////////////////////////////////////////// //
 
-	_, errs = NewArguments().Parse([]string{"--asd="}, Map{"t:test": &V{}})
+	_, errs = NewArguments().Parse([]string{"--asd="}, Map{"t:test": {}})
 
 	c.Assert(errs, Not(HasLen), 0)
 	c.Assert(errs[0].Error(), Equals, "Argument --asd has wrong format")
 
 	// //////////////////////////////////////////////////////////////////////////////// //
 
-	_, errs = NewArguments().Parse([]string{"-j="}, Map{"t:test": &V{}})
+	_, errs = NewArguments().Parse([]string{"-j="}, Map{"t:test": {}})
 
 	c.Assert(errs, Not(HasLen), 0)
 	c.Assert(errs[0].Error(), Equals, "Argument -j has wrong format")
 
 	// //////////////////////////////////////////////////////////////////////////////// //
 
-	_, errs = NewArguments().Parse([]string{}, Map{"t:test": &V{Required: true}})
+	_, errs = NewArguments().Parse([]string{}, Map{"t:test": {Required: true}})
 
 	c.Assert(errs, Not(HasLen), 0)
 	c.Assert(errs[0].Error(), Equals, "Required argument test is not set")
 
 	// //////////////////////////////////////////////////////////////////////////////// //
 
-	_, errs = NewArguments().Parse([]string{"-t"}, Map{"t:test": &V{}})
+	_, errs = NewArguments().Parse([]string{"-t"}, Map{"t:test": {}})
 
 	c.Assert(errs, Not(HasLen), 0)
 	c.Assert(errs[0].Error(), Equals, "Non-boolean argument --test is empty")
 
 	// //////////////////////////////////////////////////////////////////////////////// //
 
-	_, errs = NewArguments().Parse([]string{"--test=!"}, Map{"t:test": &V{Type: FLOAT}})
+	_, errs = NewArguments().Parse([]string{"--test=!"}, Map{"t:test": {Type: FLOAT}})
 
 	c.Assert(errs, Not(HasLen), 0)
 	c.Assert(errs[0].Error(), Equals, "Argument --test has wrong format")
 
 	// //////////////////////////////////////////////////////////////////////////////// //
 
-	_, errs = NewArguments().Parse([]string{"-t=!"}, Map{"t:test": &V{Type: INT}})
+	_, errs = NewArguments().Parse([]string{"-t=!"}, Map{"t:test": {Type: INT}})
 
 	c.Assert(errs, Not(HasLen), 0)
 	c.Assert(errs[0].Error(), Equals, "Argument --test has wrong format")
 
 	// //////////////////////////////////////////////////////////////////////////////// //
 
-	_, errs = NewArguments().Parse([]string{"--test", "!"}, Map{"t:test": &V{Type: INT}})
+	_, errs = NewArguments().Parse([]string{"--test", "!"}, Map{"t:test": {Type: INT}})
 
 	c.Assert(errs, Not(HasLen), 0)
 	c.Assert(errs[0].Error(), Equals, "Argument --test has wrong format")
@@ -338,8 +370,14 @@ func (s *ArgUtilSuite) TestParsing(c *C) {
 
 	// //////////////////////////////////////////////////////////////////////////////// //
 
-	_, errs = NewArguments().Parse([]string{}, Map{"": &V{}})
+	_, errs = NewArguments().Parse([]string{}, Map{"": {}})
 
 	c.Assert(errs, Not(HasLen), 0)
 	c.Assert(errs[0].Error(), Equals, "Some argument does not have a name")
+}
+
+func (s *ArgUtilSuite) TestMerging(c *C) {
+	c.Assert(Q(), Equals, "")
+	c.Assert(Q("test"), Equals, "test")
+	c.Assert(Q("test1", "test2"), Equals, "test1 test2")
 }